Revert "lte: Merge branch 'lte-rlc-fix-pdu-delivering', MR !30"
This reverts commit 50cfb45bd126fed8936b204c7ad02ecc1d6cba4c, reversing
changes made to ae2f33cf73256fd20419507f96705df6eb884eba.

The commit is faulty because it does not consider LTE HARQ timings.
So, in LTE, would happen that a packet is delivered even in the case that the
previous packet is under HARQ retransmission.

We will check the standard and see what does apply in this case.

Natale